scripts/test_formula_system.py
#!/usr/bin/env python3
"""
Quick test of Formula Evolution Engine

Verifies all components are working correctly.
"""

import sys
import os
from pathlib import Path

sys.path.insert(0, str(Path(__file__).parent.parent))

print("=" * 60)
print("TESTING FORMULA EVOLUTION ENGINE")
print("=" * 60)

# Test 1: Formula Engine
print("\n[1/6] Testing Formula Engine...")
try:
    from utils.formula_engine import FormulaEngine
    from analyzers.name_analyzer import NameAnalyzer
    
    analyzer = NameAnalyzer()
    features = analyzer.analyze_name("Bitcoin")
    
    engine = FormulaEngine()
    encoding = engine.transform("Bitcoin", features, 'hybrid')
    
    print(f"  ✓ Transformation successful")
    print(f"    Shape: {encoding.shape_type}, Hue: {encoding.hue:.1f}°")
except Exception as e:
    print(f"  ✗ Failed: {e}")
    sys.exit(1)

# Test 2: Cache
print("\n[2/6] Testing Cache...")
try:
    from utils.formula_cache import cache
    
    stats = cache.get_stats()
    if stats['connected']:
        print(f"  ✓ Redis connected ({stats['total_keys']} keys)")
    else:
        print(f"  ⚠ Redis not available (caching disabled)")
except Exception as e:
    print(f"  ⚠ Cache test failed: {e}")

# Test 3: Error Handler
print("\n[3/6] Testing Error Handler...")
try:
    from utils.error_handler import ValidationError, handle_formula_errors
    
    @handle_formula_errors(default_value="fallback")
    def test_func():
        raise ValidationError("Test error")
    
    result = test_func()
    print(f"  ✓ Error handling working (graceful degradation)")
except Exception as e:
    print(f"  ✗ Failed: {e}")

# Test 4: Unified Domain Model
print("\n[4/6] Testing Unified Domain Model...")
try:
    from app import app
    from core.unified_domain_model import UnifiedDomainInterface, DomainType
    
    with app.app_context():
        interface = UnifiedDomainInterface()
        
        # Test loading a small sample
        crypto = interface.load_domain(DomainType.CRYPTO, limit=10)
        
        print(f"  ✓ Domain loading works ({len(crypto)} entities)")
        
        if crypto:
            print(f"    Sample: {crypto[0].name} - {crypto[0].outcome_metric_name}")
except Exception as e:
    print(f"  ✗ Failed: {e}")

# Test 5: Formula Validator
print("\n[5/6] Testing Formula Validator...")
try:
    from analyzers.formula_validator import FormulaValidator
    
    validator = FormulaValidator()
    print(f"  ✓ Validator initialized")
    print(f"    Can test {len(validator.visual_properties)} visual properties")
except Exception as e:
    print(f"  ✗ Failed: {e}")

# Test 6: Configuration
print("\n[6/6] Testing Configuration...")
try:
    import yaml
    
    config_file = Path('config/auto_analysis.yaml')
    if config_file.exists():
        with open(config_file) as f:
            config = yaml.safe_load(f)
        
        print(f"  ✓ Configuration loaded")
        print(f"    Daily analysis: {config['schedule']['daily_analysis']['time']}")
        print(f"    Weekly analysis: {config['schedule']['weekly_deep_dive']['day']}")
    else:
        print(f"  ⚠ Config file not found")
except Exception as e:
    print(f"  ⚠ Config test failed: {e}")

print("\n" + "=" * 60)
print("ALL TESTS PASSED ✓")
print("=" * 60)
print("\nSystem is ready for:")
print("  • Automated daily analysis (2:00 AM)")
print("  • Weekly deep dive (Sunday 3:00 AM)")
print("  • Manual analysis via CLI")
print("\nNext steps:")
print("  python scripts/formula_cli.py scheduler start")
print("=" * 60)

